# api/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo List REST API",
    "description": "Spesifikasi API lengkap untuk aplikasi Todo List berbasis Go, dilengkapi dengan validasi input, status code HTTP & payload, penanganan URL slug untuk tugas, serta deskripsi detail untuk tiap endpoint.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Mendapatkan Profil Pengguna",
        "description": "Mengambil detail data pengguna yang sedang login berdasarkan JWT Token yang dikirim di HTTP Header Authorization.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Data profil berhasil ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Berhasil mengambil data profil",
                  "data": {
                    "id_users": 1,
                    "nama": "Budi Santoso",
                    "email": "budi@example.com",
                    "foto_profile": "http://localhost:8080/uploads/profiles/default.png",
                    "created_at": "2026-08-07T20:00:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT tidak ada, tidak valid, atau telah kadaluwarsa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 401,
                  "status": "fail",
                  "message": "Autentikasi gagal",
                  "data": null
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Profile"],
        "summary": "Update Nama / Upload Foto Profil",
        "description": "Memperbarui nama pengguna dan/atau mengunggah berkas foto profil baru (format JPG, PNG, max 2MB).",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/ProfileUpdateRequest" },
              "example": {
                "nama": "Budi Santoso Update",
                "foto_profile": "(pilih file .jpg/.png maks 2MB)"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil pengguna berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileResponse" },
                "example": {
                  "code": 200,
                  "status": "success",
                  "message": "Profil berhasil diperbarui",
                  "data": {
                    "id_users": 1,
                    "nama": "Budi Santoso Update",
                    "email": "budi@example.com",
                    "foto_profile": "http://localhost:8080/uploads/profiles/profile_1.jpg",
                    "updated_at": "2026-08-07T21:30:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Format berkas foto tidak didukung atau ukuran melebihi batas 2MB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                },
                "example": {
                  "code": 400,
                  "status": "fail",
                  "message": "Validasi gagal",
                  "data": null,
                  "errors": {
                    "nama": "Format nama tidak valid",
                    "foto_profile": "Format gambar harus .jpg/.png dan ukuran maksimal 2MB"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token JWT tidak ditemukan atau tidak valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "code": 401,
                  "status": "fail",
                  "message": "Autentikasi gagal",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "status", "message"],
        "properties": {
          "code": { "type": "integer", "description": "HTTP status code" },
          "status": { "type": "string", "enum": ["fail", "error"] },
          "message": {
            "type": "string",
            "description": "Ringkasan pesan error"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Selalu null pada response error"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["code", "status", "message", "errors"],
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP status code (biasanya 400)"
          },
          "status": { "type": "string", "example": "fail" },
          "message": { "type": "string", "example": "Validasi gagal" },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Selalu null pada response error"
          },
          "errors": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Map pesan error per field yang gagal validasi"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id_users": { "type": "integer", "example": 1 },
          "nama": { "type": "string", "example": "Budi Santoso" },
          "email": {
            "type": "string",
            "format": "email",
            "example": "budi@example.com"
          },
          "foto_profile": {
            "type": "string",
            "nullable": true,
            "description": "URL foto profil"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "example": 200 },
          "status": { "type": "string", "example": "success" },
          "message": {
            "type": "string",
            "example": "Berhasil mengambil data profil"
          },
          "data": { "$ref": "#/components/schemas/User" }
        }
      },
...
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Budi Santoso Update",
            "description": "Nama baru pengguna (maksimal 100 karakter)"
          },
          "foto_profile": {
            "type": "string",
            "format": "binary",
            "description": "File foto profil (format .jpg, .png, maksimal 2MB)"
          }
        }
      },
...
    }
  }
}